+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Translations: [简体中文](CHANGELOG_zh.md)
 other:
 
 * depend: Upgrade Sketch 3.3.0-beta02
+* new: Added `zoomimage-compose-coil-base` and `zoomimage-compose-sketch-core` modules, which depend
+  on the non-singleton modules of Coil and Sketch respectively.
 
 ## 1.0.0-beta07